Implement a method to authenticate using a username and password
This adds a new method, kerberos.authUserKrb5Password, which
simply accepts a username and password and validates them against
the KDC. It will also validate that the KDC providing the response
is valid by matching a service key obtained from the KDC against
one stored in a local keytab.  KDC validation can be bypassed but
this is not recommended.