fix: avoid strncpy-related oob reads
`strncpy` will not 0-terminate strings if the terminator does not
fit into the target buffer. This is already handled manually in
some places, but not all; this commit rectifies that.
Additionally, if truncation does take place, we should consider
the truncated length the length of the resulting string, not the
original length.