chore: address TODO for managing state through smart pointers
Address the TODOs asking for the state to be managed through
smart pointers and as a proper class with destructors.

This is helpful for the upcoming Node-API conversion, because
the presence of C++ exceptions in Node-API (and the fact that
Node-API support also comes with multithreading support) makes
it important to use proper C++ RAII style in order to avoid
memory leaks.